PublicationsComparatorTask config extracted into a dedicated class PublicationsComparatorPlugin #84
