+++ src/main/groovy/org/mockito/release/internal/gradle/ContinuousDeliveryPlugin.java
@@ -188,23 +188,17 @@ public class ContinuousDeliveryPlugin implements Plugin<Project> {
         return TaskMaker.task(project, taskName, ReleaseNeededTask.class, new Action<ReleaseNeededTask>() {
             public void execute(final ReleaseNeededTask t) {
                 t.setDescription("Asserts that criteria for the release are met and throws exception if release not needed.");
-                t.setReleasableBranchRegex(conf.getGit().getReleasableBranchRegex());
                 t.setExplosive(true);
+                System.out.println("release needed task");
                 t.setCommitMessage(conf.getBuild().getCommitMessage());
                 t.setPullRequest(conf.getBuild().isPullRequest());
 
                 project.allprojects(new Action<Project>() {
                     public void execute(final Project subproject) {
-                        //TODO WW, let's push out the complexity of comparing publications out of BaseJavaLibraryPlugin
-                        //into a separate plugin, something like 'PublicationsComparatorPlugin'
-                        //This way, we make the plugins smaller, more fine granular, easier to reuse and comprehend
-                        //The new plugin depends on sources jar and pom gen task so
-                        // 'PublicationsComparatorPlugin' task should first apply 'BaseJavaLibraryPlugin'
-                        //when we do that, the code below should use "withType(PublicationsComparatorPlugin.class)"
-                        subproject.getPlugins().withType(BaseJavaLibraryPlugin.class, new Action<BaseJavaLibraryPlugin>() {
-                            public void execute(BaseJavaLibraryPlugin p) {
+                        subproject.getPlugins().withType(PublicationsComparatorPlugin.class, new Action<PublicationsComparatorPlugin>() {
+                            public void execute(PublicationsComparatorPlugin p) {
                                 // make this task depend on all comparePublications tasks
-                                Task task = subproject.getTasks().getByName(BaseJavaLibraryPlugin.COMPARE_PUBLICATIONS_TASK);
+                                Task task = subproject.getTasks().getByName(PublicationsComparatorPlugin.COMPARE_PUBLICATIONS_TASK);
 
                                 //TODO WW, removing comparing publications from the workflow for now
                                 //by commenting out below code
@@ -219,6 +213,7 @@ public class ContinuousDeliveryPlugin implements Plugin<Project> {
                     public void run() {
                         String branch = gitStatus.getBranch();
                         t.setBranch(branch);
+                        t.setReleasableBranchRegex(conf.getGit().getReleasableBranchRegex());
                     }
                 });
             }
