+++ src/main/resources/META-INF/gradle-plugins/org.mockito.mockito-release-tools.publications-comparator.properties
@@ -1,0 +1,1 @@
+implementation-class=org.mockito.release.internal.gradle.PublicationsComparatorPlugin
